# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Auth API — Login & Protect",
    "version": "1.0.0",
    "description": "Secure Express API backed by Supabase Auth. Sign up, log in, log out, and call protected routes with a Bearer JWT. Click Authorize, paste the access_token returned by /auth/login, then try the locked endpoints."
  },
  "servers": [{ "url": "http://localhost:3000" }],
  "tags": [
    { "name": "auth", "description": "Sign up, log in, log out" },
    { "name": "public", "description": "Open endpoints, no token needed" },
    { "name": "protected", "description": "Endpoints guarded by JWT verification" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste the access_token from POST /auth/login."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "example": { "error": "Access token required" }
      },
      "Credentials": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email", "example": "test@example.com" },
          "password": { "type": "string", "format": "password", "example": "password123" }
        },
        "required": ["email", "password"]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "TokenPair": {
        "type": "object",
        "properties": {
          "access_token": { "type": "string" },
          "refresh_token": { "type": "string" },
          "token_type": { "type": "string", "example": "bearer" },
          "expires_in": { "type": "integer", "example": 3600 },
          "user": { "$ref": "#/components/schemas/User" }
        }
      }
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "tags": ["auth"],
        "summary": "Create a new user account",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Credentials" } } }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "user": { "$ref": "#/components/schemas/User" } }
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid email/password",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Authenticate and receive a JWT",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Credentials" } } }
        },
        "responses": {
          "200": {
            "description": "Logged in; access and refresh tokens issued",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TokenPair" } } }
          },
          "400": {
            "description": "Missing email or password",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "401": {
            "description": "Invalid login credentials",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "429": {
            "description": "Too many failed login attempts",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["auth"],
        "summary": "Exchange a refresh token for a fresh access token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "refresh_token": { "type": "string" } },
                "required": ["refresh_token"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New token pair issued",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TokenPair" } } }
          },
          "400": {
            "description": "Missing refresh token",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "401": {
            "description": "Invalid or expired refresh token",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["auth"],
        "summary": "End the user's session",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Signed out" },
          "401": {
            "description": "Missing, malformed, or invalid token",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/public/info": {
      "get": {
        "tags": ["public"],
        "summary": "Read public, open data",
        "responses": {
          "200": {
            "description": "Public message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "message": { "type": "string" } },
                  "example": { "message": "Welcome stranger! This info is public." }
                }
              }
            }
          }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "tags": ["protected"],
        "summary": "Read the logged-in user's profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Safe user metadata",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } }
          },
          "401": {
            "description": "Missing, malformed, or invalid token",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "tags": ["protected"],
        "summary": "A second locked door reusing the same guard",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Private dashboard message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "message": { "type": "string" } }
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, or invalid token",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/protected/admin": {
      "get": {
        "tags": ["protected"],
        "summary": "Admin-only route — 403 for authenticated non-admins",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Admin content",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "message": { "type": "string" } }
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, or invalid token",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "403": {
            "description": "Authenticated but not an admin",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    }
  }
}